refactor: Extract getOperatorsFeatureCollection
