+++ src/components/map.stories.tsx
@@ -5,22 +5,21 @@ import { I18nProvider } from "@lingui/react";
 import { Box, List, ListItemButton } from "@mui/material";
 import { Decorator } from "@storybook/react";
 import * as turf from "@turf/turf";
-import { easeExpIn, mean, median, sort } from "d3";
-import {
-  Feature,
-  FeatureCollection,
-  Geometry,
-  MultiPolygon,
-  Polygon,
-} from "geojson";
-import { groupBy, keyBy } from "lodash";
+import { easeExpIn, mean, median } from "d3";
+import { Feature, Geometry, MultiPolygon, Polygon } from "geojson";
+import { keyBy } from "lodash";
 import { useCallback, useMemo, useRef, useState } from "react";
 import { ObjectInspector } from "react-inspector";
 import { createClient, Provider } from "urql";
 
 import { ChoroplethMap } from "src/components/map";
 import { getFillColor, getZoomFromBounds } from "src/components/map-helpers";
-import { MunicipalityFeatureCollection, useGeoData } from "src/data/geo";
+import {
+  getOperatorsFeatureCollection,
+  MunicipalityFeatureCollection,
+  OperatorLayerProperties,
+  useGeoData,
+} from "src/data/geo";
 import { useFetch } from "src/data/use-fetch";
 import { useColorScale } from "src/domain/data";
 import { useSunshineTariffQuery } from "src/graphql/queries";
@@ -29,7 +28,6 @@ import { truthy } from "src/lib/truthy";
 import {
   ElectricityCategory,
   getOperatorMunicipalities,
-  OperatorMunicipalityRecord,
 } from "src/rdf/queries";
 
 import { props } from "./map.mock";
@@ -69,102 +67,6 @@ export const Example = () => {
       </Box>
     </I18nProvider>
   );
-};
-
-type OperatorLayerProperties = {
-  municipalityCount: number;
-  operators: number[];
-};
-
-const multiGroupBy = <T,>(arr: T[], iter: (item: T) => string[]) => {
-  const res: Record<string, T[]> = {};
-  for (let i = 0; i < arr.length; i++) {
-    const item = arr[i];
-    const keys = iter(item);
-    for (let j = 0; j < keys.length; j++) {
-      const key = keys[j];
-      if (!res[key]) {
-        res[key] = [];
-      }
-      res[key].push(item);
-    }
-  }
-  return res;
-};
-
-const getOperatorsFeatureCollection = (
-  operatorMunicipalities: OperatorMunicipalityRecord[],
-  municipalities: MunicipalityFeatureCollection
-): FeatureCollection<
-  MultiPolygon | Polygon,
-  OperatorLayerProperties
-> | null => {
-  if (!operatorMunicipalities || !municipalities) return null;
-
-  const operatorsByMunicipality = groupBy(
-    operatorMunicipalities,
-    "municipality"
-  );
-  const municipalitySet = Array.from(
-    new Set(operatorMunicipalities.map((x) => x.municipality))
-  );
-
-  // Group municipalities by operator, a municipality will be part of several groups if it
-  // has several operators, the groups are sorted so that the multiple operators are last
-  // Example: Kilchberg is served by EWL and EWA, it will be part of 3 groups
-  // EWL, EWA, EWL/EWA
-  const municipalitiesByOperators = multiGroupBy(municipalitySet, (x) => {
-    const operatorIds = operatorsByMunicipality[x].map((x) => x.operator);
-    const all = sort(operatorIds).join("/");
-    return [...operatorIds, all];
-  });
-
-  const municipalitiesById = keyBy(municipalities.features, "id");
-  const operatorFeatures = Object.entries(municipalitiesByOperators)
-    .map(([operators, municipalities]) => {
-      // Get geometry features for all municipalities of this operator
-      const municipalityFeatures = municipalities
-        .map((muni) => municipalitiesById[muni])
-        .filter(Boolean);
-
-      if (municipalityFeatures.length === 0) {
-        console.warn(
-          `No geometry found for operator ${operators} with municipalities: ${municipalities
-            .map((m) => m)
-            .join(", ")}`
-        );
-        return null;
-      }
-      const featureCollection = turf.featureCollection(
-        municipalityFeatures.map((feat) =>
-          turf.feature(feat.geometry as Polygon | MultiPolygon)
-        )
-      );
-
-      const geometry =
-        municipalityFeatures.length > 1
-          ? turf.union(featureCollection)?.geometry
-          : featureCollection.features[0].geometry;
-
-      if (!geometry) {
-        return null;
-      }
-      return {
-        type: "Feature" as const,
-        properties: {
-          operators: operators.split("/").map((x) => parseInt(x, 10)),
-          municipalityCount: municipalities.length,
-        },
-        geometry: geometry,
-      };
-    })
-    .filter(truthy);
-
-  // Create the final GeoJSON
-  return {
-    type: "FeatureCollection",
-    features: operatorFeatures,
-  };
 };
 
 const sunshineAttributeToElectricityCategory: Partial<
@@ -384,7 +286,7 @@ export const Operators = () => {
                   },
                 }),
 
-                // Transparent layers for hover effect
+                // Transparent layer only used for hover effect
                 new GeoJsonLayer<
                   Feature<Polygon | MultiPolygon, OperatorLayerProperties>
                 >({
@@ -407,9 +309,7 @@ export const Operators = () => {
                 new GeoJsonLayer({
                   id: "municipality-layer",
                   data: geoData.municipalities.features,
-                  filled: true,
                   getLineColor: [255, 255, 255],
-                  getFillColor: TRANSPARENT,
                   highlightColor: [0, 0, 255],
                   lineWidthUnits: "pixels",
                   stroked: true,
